backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
apscheduler
httpx
feedparser
tenacity
groq
celery
redis
scikit-learn
numpy
scipy
sentence-transformers